Switch pixel position to float

+++ src/main/scala/org/kneelawk/imagegenerator/ImageSink.scala
@@ -1,5 +1,5 @@
 package org.kneelawk.imagegenerator
 
 trait ImageSink {
-  def apply(x: Int, y: Int, pixel: Color)
+  def apply(x: Float, y: Float, pixel: Color)
 }
